requirements.txt
numpy
opencv-python==3.4.2.17
matplotlib
pyserial
websockets
websocket_client
pillow
cryptography
drawnow
Cython
